Implement ``StructSerializableNode`` & ``StructSerializableMapping``
This is unfortunately slower than ``PickleSerializableMapping``, mostly
due to the need to serialize variable-length strings. It also cannot
currently handle storing numbers over 2^64-1 (18446744073709551615), as
they are stored as an unsigned long long -- this means that storing a
deep tree or a tree with large files or costs is impossible with this
class.